Fix 4 compiler bugs discovered during showcase demo creation
Bug 1: Top-level const/let declarations no longer hang the compiler.
Added ItemKind::Const(VarDecl) to the AST, parser support for const/let
as module-level items (including export const), RustItem::Const with
RustConstItem in the IR, and emitter support. Also fixed the infinite
loop in parse_item's fallback — the synchronize() method would stop on
const/let tokens without advancing, causing an infinite loop. The
fallback now advances past the unexpected token before synchronizing.

Bug 2: Struct literals inside array expressions now resolve the correct
type name. extract_named_type() for Vec<Todo> was returning "Vec"
instead of "Todo". Now extracts the element type name for collection
types (Vec, HashSet), so const todos: Array<Todo> = [{...}] correctly
resolves struct literals as Todo{...} instead of Vec{...}.

Bug 3: Block-body closures in .reduce() now emit proper block bodies
instead of a _ placeholder. Changed IteratorTerminal::Fold::body from
Box<RustExpr> to RustClosureBody, and updated lower_array_reduce,
merge_reduce_into_chain, and the emitter to preserve and emit block
closure bodies.

Bug 4: .length now emits .len() as i64 instead of bare .len(), matching
RustScript's default numeric type and preventing usize/i64 type
mismatches. Added RustExprKind::Cast variant to the IR and emitter.

+++ crates/rsc-driver/tests/phase2_snapshots.rs
@@ -267,7 +267,7 @@ async function main() {
 #[tokio::main]
 async fn main() {
     let names = getNames().await;
-    println!("{}", names.len());
+    println!("{}", names.len() as i64);
 }
 "#;
 
@@ -435,7 +435,7 @@ fn test_snapshot_p2_string_split_method() {
 
     let expected = r#"fn main() {
     let parts = "a,b,c".to_string().split(",").map(|s| s.to_string()).collect::<Vec<String>>();
-    println!("{}", parts.len());
+    println!("{}", parts.len() as i64);
 }
 "#;
 
@@ -586,7 +586,7 @@ fn test_snapshot_p2_string_length_property() {
 
     let expected = r#"fn main() {
     let name = "Alice".to_string();
-    let len = name.len();
+    let len = name.len() as i64;
     println!("{}", len);
 }
 "#;
