+++ crates/rsc-driver/tests/snapshots.rs
@@ -2118,7 +2118,7 @@ fn test_snapshot_string_split_method() {
 
     let expected = r#"fn main() {
     let parts = "a,b,c".to_string().split(",").map(|s| s.to_string()).collect::<Vec<String>>();
-    println!("{}", parts.len());
+    println!("{}", parts.len() as i64);
 }
 "#;
 
@@ -2197,7 +2197,7 @@ fn test_snapshot_string_length_property() {
 
     let expected = r#"fn main() {
     let name = "Alice".to_string();
-    let len = name.len();
+    let len = name.len() as i64;
     println!("{}", len);
 }
 "#;
@@ -2395,7 +2395,7 @@ function main(): void {
 
     let expected = "\
 fn len(s: &str) -> i64 {
-    return s.len();
+    return s.len() as i64;
 }
 
 fn main() {
@@ -2459,7 +2459,7 @@ function main(): void {
 
     let expected = "\
 fn countItems(items: &Vec<i32>) -> i64 {
-    return items.len();
+    return items.len() as i64;
 }
 
 fn main() {
